% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecuador}
\alias{ecuador}
\title{Ecuador Dollarization}
\format{data.frame}
\usage{
data(ecuador)
}
\description{
A \code{data.frame} containing data across 41 countries. including Ecuador.
}
\keyword{datasets}
