% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth}
\alias{synth}
\title{Synthetic Control Method}
\usage{
synth(data, y, treat, time, group, ..., op = mean, method = "Nelder-Mead")
}
\arguments{
\item{data}{data.frame}

\item{y}{character, symbol, or number denoting the column containing the dependent variable}

\item{treat}{character, symbol, or number denoting the column containing the treatment dummy}

\item{time}{character, symbol, or number denoting the column containing the time variable}

\item{group}{character, symbol, or number denoting the column containing the grouping variable}

\item{...}{characters, symbols, or numbers denoting the columns containing co-variates}

\item{op}{operation used on X0 for determining weights. Default is 'mean'}

\item{method}{optimization method accepted by \code{optim}. Default is "Nelder-Mead"}
}
\value{
list of tables containing \code{y}, the estimated counterfactual, weights, plots, the regression results
}
\description{
Applies the synthetic control method to a variable within a data.frame. The user must specify
which variable receives the treatment, a dummy variable taking a value of '0' before treatment
and '1' after the treatment, a variable specifying groups, and the co-variates used to
determine weights.
}
\examples{
data(ecuador)
results <- synth(
  pwt_mod,                                         # data
  GDPpc, Treat, year, country,                     # y, treat, time, group
  pop, PR, KL, xr, delta, ex, im, s, Pex, Pim, Pk  # co-variates
)

# Plot Results
results$plot_path
results$plot_gap
results$plot_jitter

# Show Weights
results$w_group
results$w_pred

# Show Differences before and after
results$difference

# Regression summary of Treatment on the Gap
summary(results$lm)

# Data.frame with Treated and Synthetic Series
results$results

}
