% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth_many}
\alias{synth_many}
\title{Apply the Synthetic Control Method to Multiple Treated Groups}
\usage{
synth_many(data, y, treat, time, group, ..., op = mean, method = "Nelder-Mead")
}
\arguments{
\item{data}{data.frame}

\item{y}{character, symbol, or number denoting the column containing the dependent variable}

\item{treat}{character, symbol, or number denoting the column containing the treatment dummy}

\item{time}{character, symbol, or number denoting the column containing the time variable}

\item{group}{character, symbol, or number denoting the column containing the grouping variable}

\item{...}{characters, symbols, or numbers denoting the columns containing co-variates}

\item{op}{operation used on X0 for determining weights. Default is 'mean'}

\item{method}{optimization method accepted by \code{optim}. Default is "Nelder-Mead"}
}
\value{
list of tables containing \code{y}, the estimated counterfactual, weights, plots, the regression results
}
\description{
Applies the synthetic control method to multiple treatment groups within a data.frame. See
\code{synth} for the case with only one group receiving treatment.
}
\examples{

}
